requirements.txt
ultralytics>=8.3.0
pyyaml>=6.0
opencv-python>=4.9.0
python-dotenv>=1.0.1
openai>=1.51.0
django>=5.0
requests>=2.31.0
